requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0
django-stdimage==5.3.0
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
sqlparse==0.4.2
static3==0.7.0
tzdata==2021.5
